Add parser capable of handling all the main matter
This parser implementation tries to handle all[1] commands generically
with a single AST node: Atom::Command.

That has proven to be a bad design, since it is impossible to
_generically_ differentiate a command followed by a group from a command
with a single required argument.

After the necessary rewrite, the parser should also be made to correctly
allow/disallow NEWLINEs between commands and their arguments.

Another possible issue with this parser is that implicit_par only
handles true empty lines, not _visibly_ empty ones; I cannot remember
which of those TeX follows.  Additionally, it might be a good idea to
have a implicit_par at EOI.

[1] All commands apart from \include{<path>}.